requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
commonmark==0.9.1
dictdiffer==0.8.1
google-api-core==1.26.0
google-auth==1.27.0
google-cloud-core==1.6.0
google-cloud-translate==3.0.2
googleapis-common-protos==1.53.0
grpcio==1.36.0
idna==2.10
libcst==0.3.17
mypy-extensions==0.4.3
notion==0.0.28
packaging==20.9
proto-plus==1.14.2
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-slugify==4.0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
soupsieve==2.2
text-unidecode==1.3
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==2.1
urllib3==1.26.3
